% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_abstract.R
\name{add_abstract}
\alias{add_abstract}
\title{Add an abstract file to a Quarto report}
\usage{
add_abstract(report_path)
}
\arguments{
\item{report_path}{The path to the folder containing the report.}
}
\value{
The name of the index file.
}
\description{
This function adds an abstract file to a Quarto report.
}
