% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_address.R
\name{add_address}
\alias{add_address}
\title{Format an address in YAML}
\usage{
add_address(type = "client")
}
\arguments{
\item{type}{The type of address.
Defaults to \code{"client"}.}
}
\value{
A character vector containing the formatted address.
}
\description{
Format an address in YAML
}
