% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_author.R
\name{add_author}
\alias{add_author}
\title{Add an author or reviewer to the \verb{_quarto.yml} file}
\usage{
add_author(report_path = ".", reviewer = FALSE)
}
\arguments{
\item{report_path}{The path to the folder containing the report.
Defaults to the current working directory.}

\item{reviewer}{If \code{TRUE}, the person is added as a reviewer.
Defaults to \code{FALSE}.
If \code{FALSE}, the person is added as an author.}
}
\value{
The path to the \verb{_quarto.yml} file.
}
\description{
Add an author or reviewer to the \verb{_quarto.yml} file
}
