% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bibliography.R
\name{add_bibliography}
\alias{add_bibliography}
\title{Add a bibliography file to a Quarto report or website}
\usage{
add_bibliography(report_path = ".")
}
\arguments{
\item{report_path}{The path to the folder containing the report.
Defaults to the current working directory.}
}
\value{
The name of the bibliography file.
}
\description{
This function adds a bibliography file to a Quarto report or website.
}
