% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chapter.R
\name{add_chapter}
\alias{add_chapter}
\title{Add a chapter file to a Quarto report}
\usage{
add_chapter(report_path, title, filename, toc = TRUE)
}
\arguments{
\item{report_path}{The path to the folder containing the report.}
}
\value{
The name of the chapter file.
}
\description{
This function adds a chapter file to a Quarto report.
}
