% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cover.R
\name{add_cover}
\alias{add_cover}
\title{Add cover to \verb{_quarto.yml}}
\usage{
add_cover(report_path, cover_pdf)
}
\arguments{
\item{report_path}{The path to the folder containing the report.}

\item{cover_pdf}{The path to a PDF file.
The first page of this PDF file will be used as the cover of the report.}
}
\description{
Add cover to \verb{_quarto.yml}
}
