% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_index}
\alias{add_index}
\title{Add an index file to a Quarto report or website}
\usage{
add_index(report_path = ".")
}
\arguments{
\item{report_path}{The path to the folder containing the report.
Defaults to the current working directory.}
}
\value{
The name of the index file.
}
\description{
This function adds an index file to a Quarto report or website.
}
