% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_intro.R
\name{add_intro}
\alias{add_intro}
\title{Add an introduction file to a Quarto report}
\usage{
add_intro(report_path)
}
\arguments{
\item{report_path}{The path to the folder containing the report.}
}
\value{
The name of the index file.
}
\description{
This function adds an introduction file to a Quarto report.
}
