% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_recommendations.R
\name{add_recommendations}
\alias{add_recommendations}
\title{Add a recommendations section to a \code{flandersqmd} report}
\usage{
add_recommendations(report_path, lof = TRUE, lot = TRUE)
}
\arguments{
\item{report_path}{The path to the folder containing the report.}

\item{lof}{A logical value indicating whether to add a list of figures.
Defaults to \code{TRUE}.
If \code{TRUE}, a list of figures is added to the pdf version of the report.}

\item{lot}{A logical value indicating whether to add a list of tables.
Defaults to \code{TRUE}.
If \code{TRUE}, a list of tables is added to the pdf version of the report.}

\item{lang}{The language of the report.
Defaults to \code{"nl-BE"}.
The language must be one of the following:
\code{"nl-BE"}, \code{"en-GB"} or \code{"fr-FR"}.}
}
\description{
This function adds a recommendations section to a \code{flandersqmd} report.
The file also add a table of contents, a list of figures and a list of
tables to the pdf version of the report.
}
