% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author2yaml.R
\name{author2yaml}
\alias{author2yaml}
\title{Convert a data frame with author information to YAML format}
\usage{
author2yaml(author, corresponding = FALSE)
}
\arguments{
\item{author}{a data frame with author information.
It should contain the columns:
\itemize{
\item \code{given}: the given name of the author
\item \code{family}: the family name of the author
\item \code{email}: the email address of the author (optional)
\item \code{orcid}: the ORCID of the author (optional)
\item \code{affiliation}: the affiliation of the author (optional)
}}

\item{corresponding}{a logical value indicating whether the author is the
corresponding author.
If \code{TRUE}, the email address of the author must be provided.}
}
\value{
a character vector containing the YAML representation of the author
}
\description{
Convert a data frame with author information to YAML format
}
