% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_website.R
\name{create_website}
\alias{create_website}
\title{Create a template for a \code{flandersqmd} website}
\usage{
create_website(path = ".", websitename, version = "main")
}
\arguments{
\item{path}{The folder in which to create the folder containing the website.
Defaults to the current working directory.
It also creates an RStudio project file in the website folder.
When ran from RStudio, the project will be opened automatically in a new
session.}

\item{websitename}{The folder name of the website.
The location of the folder \code{websitename} depends on the content of \code{path}.
When \code{path} is a \code{checklist::checklist} project, you will find the new
website at \code{path/source/websitename}.
When \code{path} is a \code{checklist::checklist} package, you will find the new
website at \code{path/inst/websitename}.
Otherwise you will find the new website at \code{path/websitename}.}

\item{version}{The version of the \code{flandersqmd-website} extension to use.
Defaults to \code{"main"}, which refers to the current version.}
}
\description{
Create a template for a \code{flandersqmd} website
}
\seealso{
Other utils:
\code{\link[=create_report]{create_report()}},
\code{\link[=inbo_website]{inbo_website()}}
}
\concept{utils}
