% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branches_tags.R
\name{get_branches_tags}
\alias{get_branches_tags}
\title{Get branches and tags of a GitHub repository}
\usage{
get_branches_tags(owner, repo)
}
\arguments{
\item{owner}{Repository owner}

\item{repo}{Repository name}
}
\value{
A sorted character vector of branch and tag names, excluding the
\code{gh-pages} branch.
}
\description{
Get branches and tags of a GitHub repository
}
\concept{git}
