% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_website.R
\name{inbo_website}
\alias{inbo_website}
\title{Render the report into a zip file for the INBO website}
\usage{
inbo_website(path = ".")
}
\arguments{
\item{path}{The path to the directory where the report is located}
}
\description{
Render the report into a zip file for the INBO website
}
\seealso{
Other utils:
\code{\link[=create_report]{create_report()}}
}
\concept{utils}
