% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_missing_metadata.R
\name{insert_missing_metadata}
\alias{insert_missing_metadata}
\title{Insert missing metadata in \verb{_quarto.yml}}
\usage{
insert_missing_metadata(report_path)
}
\arguments{
\item{report_path}{The path to the folder containing the report.}
}
\description{
Insert missing metadata in \verb{_quarto.yml}
}
