% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chapter.R
\name{add_chapter}
\alias{add_chapter}
\title{Add a chapter file to a Quarto report}
\usage{
add_chapter(report_path, title, filename, toc = TRUE)
}
\arguments{
\item{report_path}{The path to the folder containing the report.}

\item{title}{The title of the chapter.
If missing, the chapter is assumed the introduction with a default title
based on the language.}

\item{filename}{The name of the chapter file.
If missing, the chapter is assumed the introduction with a default filename
based on the language.}

\item{toc}{A logical value indicating whether to add a local table of
contents.
Defaults to \code{TRUE}.}
}
\value{
The name of the chapter file.
}
\description{
This function adds a chapter file to a Quarto report.
}
