% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a template for a \code{flandersqmd} report}
\usage{
create_report(path = ".", reportname, version = "main", shortname)
}
\arguments{
\item{path}{The folder in which to create the folder containing the report.
Defaults to the current working directory.
It also creates an RStudio project file in the report folder.
When ran from RStudio, the project will be opened automatically in a new
session.}

\item{reportname}{The folder name of the report.
The location of the folder \code{reportname} depends on the content of \code{path}.
When \code{path} is a \code{checklist::checklist} project, you will find the new report
at \code{path/source/reportname}.
When \code{path} is a \code{checklist::checklist} package, you will find the new report
at \code{path/inst/reportname}.
Otherwise you will find the new report at \code{path/reportname}.}

\item{version}{The version of the \code{flandersqmd-book} extension to use.
Defaults to \code{"main"}, which refers to the current version.}

\item{shortname}{Deprecated.
Use \code{reportname} instead.}
}
\description{
Create a template for a \code{flandersqmd} report
}
\seealso{
Other utils:
\code{\link[=inbo_website]{inbo_website()}}
}
\concept{utils}
